Corrected FormDate

+++ src/main/java/ru/netology/service/FormDate.java
@@ -1,7 +1,7 @@
 package ru.netology.service;
 
 public class FormDate {
-    public long day;
-    public long month;
-    public long year;
+    public int day;
+    public int month;
+    public int year;
 }
